[wemo] refactor duplicate code part2
